last .ver

+++ temp_watch.py
@@ -2,9 +2,12 @@ import os
 import csv
 import time
 
-with open('temp_record.csv', 'w') as f:
-    writer = csv.writer(f, delimiter=',')
-    writer.writerow(['time', 'temp', 'RAM'])
+if os.path.isfile('temp_record.csv'):
+	pass
+else:
+	with open('temp_record.csv', 'w') as f:
+    		writer = csv.writer(f, delimiter=',')
+    		writer.writerow(['time', 'temp', 'RAM'])
 
 
 # Return RAM information (unit=kb) in a list
